Added basic support for class level constraints
Added support for cascaded classes with the @Valid annotation

BeanDescriptor, ElementDescriptor, and PropertyDescriptor are now 100%

PropertyDescriptor is about 33% still

Added a ConstraintDescriptor decorator.  This will cause memory use (but not code generation) to go up some but should, otherwise, 
prevent problems where the various java.util.Sets in the application see a constraint definition (ConstraintDescription) on two 
seperate properties as one and the same.  This can cause some bizzare issues in the counting process.  This should be fixed going
forward.