Fixed a ton of confusion in the way that the ConstraintFinder was implemented.  The code is a lot less problematic and is a good bit easier to understand.  The recursion could be implemented better but it works for now.
Implemented the ability to search declaredOn with the ConstraintFinder.  For now the compiled side is missing an implementation of IReflector.declaredOn(...) so that part is not complete but the stub is there.  This brings the ConstraintFinder implementation to:
Reflective:
    Bean: 90% (ElementType.TYPE is not included)
    Property: 100%
Compiled:
    Bean: 66%
    Property: 66%

The first tests for the validators were added as a proof of concept.  The NotNullObject valdiator is complete.  It initializes properly and validates properly.  Consider this a first milestone.