rearranged the test suite structure to make it seperate from the test cases
added a NotEmpty validator (NotNull, Size > 1) that works on a Collections, Arrays, Strings, and Maps

had to fix a lot of problems with the way that validators are resolved, especially on composed constraints.  basically it was a modification to the way that caching is maintained.

ValidatorImpl is now covered around 98%