Fixed a few classes and problem areas as well as stitched together more of the configuration classes.  ValidationFactoryImpl probably should not continue to be a singleton... and neither should some of the other classes at that level.
Began to work with the JSR-303 TCK and have a basic working pom/project for it.  At this checkin it is only 31% compatible with the TCK and there are some infinite loop problems as well.  Now trying to reproduce them and (in an effort to do so) will probably end up porting some of the test cases over.