updated to a more "fair" representaiton of the license.
the old license text referecned an LLC (which isn't doing this work) and only me.  it should reference the community at large and the gwt-validation project.

the license text from before was just a holdover from when I was doing this on my own time.  my appologies to everyone who has helped.