unit tests and backbone behind issues #43 and #44.
removed some of the more complicated aspects around the Reflector implementations for client side.  It no longer is a parameterized type to reduce complexity.  This is so that we can chain getValues() and use getReflector recursively.  

parts of issue #44 are fixed but not the crucial part: the anonymous or inner type support.