in preparation for fixing some of the tck problems and 2.0 defects we've cleaned up the way that test cases are created.
instead of creating a static test and one test case caller for each, now we create a base test case and then extend it.  this abuses the GWTTestCase.getModuleName() method.

i think, in the future, there is a way we can use reflection or method interception to take this even further, or possibly a clever use of GWTTestCase.isPureJava().

Either way, this way takes less code and should be easier to maintain in the log run.